Refactor negotiation stream parts into focused modules
