+++ crates/transport/src/negotiation/parts/stream_parts/copy.rs
@@ -1,0 +1,182 @@
+use std::collections::TryReserveError;
+use std::io::{self, IoSliceMut, Write};
+
+use super::super::super::{BufferedCopyTooSmall, NegotiationBufferAccess};
+use super::NegotiatedStreamParts;
+
+impl<R> NegotiatedStreamParts<R> {
+    /// Copies the buffered negotiation data into a caller-provided vector without consuming it.
+    ///
+    /// The helper mirrors [`Self::buffered`] but writes the bytes into an owned [`Vec<u8>`], making
+    /// it straightforward to persist handshake transcripts or reuse heap storage across sessions.
+    /// The vector is cleared before the data is appended. Any additional capacity required to
+    /// complete the copy is reserved using [`Vec::try_reserve`], with allocation failures reported
+    /// via [`TryReserveError`].
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use rsync_transport::sniff_negotiation_stream;
+    /// use std::io::Cursor;
+    ///
+    /// let parts = sniff_negotiation_stream(Cursor::new(b"@RSYNCD: 31.0\nreply".to_vec()))
+    ///     .expect("sniff succeeds")
+    ///     .into_parts();
+    /// let mut replay = Vec::with_capacity(4);
+    /// parts
+    ///     .copy_buffered_into_vec(&mut replay)
+    ///     .expect("vector can reserve space for replay bytes");
+    /// assert_eq!(replay.as_slice(), parts.buffered());
+    /// ```
+    #[must_use = "the result reports whether the replay vector had sufficient capacity"]
+    pub fn copy_buffered_into_vec(&self, target: &mut Vec<u8>) -> Result<usize, TryReserveError> {
+        NegotiationBufferAccess::copy_buffered_into_vec(self, target)
+    }
+
+    /// Copies the unread portion of the buffered negotiation data into a caller-provided vector without consuming it.
+    #[must_use = "the result reports whether the replay vector had sufficient capacity"]
+    pub fn copy_buffered_remaining_into_vec(
+        &self,
+        target: &mut Vec<u8>,
+    ) -> Result<usize, TryReserveError> {
+        NegotiationBufferAccess::copy_buffered_remaining_into_vec(self, target)
+    }
+
+    /// Appends the buffered negotiation data to a caller-provided vector without consuming it.
+    ///
+    /// The helper mirrors [`NegotiatedStream::extend_buffered_into_vec`] but operates on decomposed
+    /// stream parts. Callers that temporarily separate the components can therefore continue to
+    /// accumulate handshake transcripts inside pre-existing log buffers. Additional capacity is
+    /// reserved via [`Vec::try_reserve`]; on success the appended length matches the buffered
+    /// payload and the replay cursor remains untouched.
+    #[must_use = "the result reports whether additional capacity was successfully reserved"]
+    pub fn extend_buffered_into_vec(&self, target: &mut Vec<u8>) -> Result<usize, TryReserveError> {
+        NegotiationBufferAccess::extend_buffered_into_vec(self, target)
+    }
+
+    /// Appends the unread portion of the buffered negotiation transcript to `target` without consuming it.
+    #[must_use = "the result reports whether additional capacity was successfully reserved"]
+    pub fn extend_buffered_remaining_into_vec(
+        &self,
+        target: &mut Vec<u8>,
+    ) -> Result<usize, TryReserveError> {
+        NegotiationBufferAccess::extend_buffered_remaining_into_vec(self, target)
+    }
+
+    /// Copies the buffered negotiation bytes into the destination vector without consuming them.
+    ///
+    /// This mirrors [`NegotiatedStream::copy_buffered_into`], allowing callers that temporarily
+    /// decompose the stream into parts to observe the sniffed prefix and remainder while preserving
+    /// the replay state. The destination is cleared before data is appended; if additional capacity
+    /// is required a [`TryReserveError`] is returned and the original contents remain untouched.
+    #[must_use = "the returned length reports how many bytes were copied and whether allocation succeeded"]
+    pub fn copy_buffered_into(&self, target: &mut Vec<u8>) -> Result<usize, TryReserveError> {
+        NegotiationBufferAccess::copy_buffered_into(self, target)
+    }
+
+    /// Copies the buffered negotiation data into the provided vectored buffers without consuming it.
+    ///
+    /// The helper mirrors [`Self::copy_buffered_into_slice`] while operating on a mutable slice of
+    /// [`IoSliceMut`]. This is useful when the stream has been decomposed into parts but callers
+    /// still need to scatter the sniffed negotiation transcript across multiple scratch buffers
+    /// without cloning the stored bytes.
+    ///
+    /// # Errors
+    ///
+    /// Returns [`BufferedCopyTooSmall`] if the combined capacity of `bufs` is smaller than the
+    /// buffered negotiation payload.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use rsync_transport::sniff_negotiation_stream;
+    /// use std::io::{Cursor, IoSliceMut};
+    ///
+    /// let parts = sniff_negotiation_stream(Cursor::new(b"@RSYNCD: 31.0\nreply".to_vec()))
+    ///     .expect("sniff succeeds")
+    ///     .into_parts();
+    /// let expected = parts.buffered().to_vec();
+    /// let mut first = [0u8; 10];
+    /// let mut second = [0u8; 32];
+    /// let mut bufs = [IoSliceMut::new(&mut first), IoSliceMut::new(&mut second)];
+    /// let copied = parts
+    ///     .copy_buffered_into_vectored(&mut bufs)
+    ///     .expect("buffers are large enough");
+    ///
+    /// let prefix_len = first.len().min(copied);
+    /// let mut assembled = Vec::new();
+    /// assembled.extend_from_slice(&first[..prefix_len]);
+    /// let remainder_len = copied - prefix_len;
+    /// if remainder_len > 0 {
+    ///     assembled.extend_from_slice(&second[..remainder_len]);
+    /// }
+    /// assert_eq!(assembled, expected);
+    /// ```
+    #[must_use = "the return value conveys whether the provided buffers were large enough"]
+    pub fn copy_buffered_into_vectored(
+        &self,
+        bufs: &mut [IoSliceMut<'_>],
+    ) -> Result<usize, BufferedCopyTooSmall> {
+        NegotiationBufferAccess::copy_buffered_into_vectored(self, bufs)
+    }
+
+    /// Copies the buffered negotiation data into the caller-provided slice without consuming it.
+    #[must_use = "the result indicates if the destination slice could hold the buffered bytes"]
+    pub fn copy_buffered_into_slice(
+        &self,
+        target: &mut [u8],
+    ) -> Result<usize, BufferedCopyTooSmall> {
+        NegotiationBufferAccess::copy_buffered_into_slice(self, target)
+    }
+
+    /// Copies the unread portion of the buffered negotiation data into `target` without consuming it.
+    #[must_use = "the result indicates if the destination slice could hold the remaining buffered bytes"]
+    pub fn copy_buffered_remaining_into_slice(
+        &self,
+        target: &mut [u8],
+    ) -> Result<usize, BufferedCopyTooSmall> {
+        NegotiationBufferAccess::copy_buffered_remaining_into_slice(self, target)
+    }
+
+    /// Copies the buffered negotiation data into a caller-provided array without consuming it.
+    #[must_use = "the result indicates if the destination array could hold the buffered bytes"]
+    pub fn copy_buffered_into_array<const N: usize>(
+        &self,
+        target: &mut [u8; N],
+    ) -> Result<usize, BufferedCopyTooSmall> {
+        NegotiationBufferAccess::copy_buffered_into_array(self, target)
+    }
+
+    /// Copies the unread portion of the buffered negotiation data into a caller-provided array without consuming it.
+    #[must_use = "the result indicates if the destination array could hold the remaining buffered bytes"]
+    pub fn copy_buffered_remaining_into_array<const N: usize>(
+        &self,
+        target: &mut [u8; N],
+    ) -> Result<usize, BufferedCopyTooSmall> {
+        NegotiationBufferAccess::copy_buffered_remaining_into_array(self, target)
+    }
+
+    /// Streams the buffered negotiation data into the provided writer without consuming it.
+    #[must_use = "the returned length reports how many bytes were written and surfaces I/O failures"]
+    pub fn copy_buffered_into_writer<W: Write>(&self, target: &mut W) -> io::Result<usize> {
+        NegotiationBufferAccess::copy_buffered_into_writer(self, target)
+    }
+
+    /// Streams the unread portion of the buffered negotiation data into the provided writer.
+    #[must_use = "the returned length reports how many bytes were written and surfaces I/O failures"]
+    pub fn copy_buffered_remaining_into_writer<W: Write>(
+        &self,
+        target: &mut W,
+    ) -> io::Result<usize> {
+        NegotiationBufferAccess::copy_buffered_remaining_into_writer(self, target)
+    }
+
+    /// Copies the unread portion of the buffered negotiation data into the provided vectored buffers without consuming it.
+    #[must_use = "the return value conveys whether the provided buffers were large enough"]
+    pub fn copy_buffered_remaining_into_vectored(
+        &self,
+        bufs: &mut [IoSliceMut<'_>],
+    ) -> Result<usize, BufferedCopyTooSmall> {
+        NegotiationBufferAccess::copy_buffered_remaining_into_vectored(self, bufs)
+    }
+}
